Improve class names
