+++ tools/javadoc-style/src/test/java/io/spine/tools/javadoc/style/formatting/RemovePreTagsTest.java
@@ -26,43 +26,41 @@
 
 package io.spine.tools.javadoc.style.formatting;
 
-import com.google.common.base.Joiner;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import java.util.Collections;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import static com.google.common.truth.Truth.assertThat;
-import static java.lang.System.lineSeparator;
+@DisplayName("PreTagFormatting should")
+class RemovePreTagsTest {
 
-@DisplayName("`LineFormatting` should")
-class LineFormattingTest {
+    private static final String RAW_PROTO_DOC =
+            "/** <pre> Doc goes here </pre> <code>string field = 1;</code> */";
+    private static final String PROCESSED_PROTO_DOC =
+            "/**  Doc goes here  <code>string field = 1;</code> */";
 
-    private final Formatting formatting = new NothingFormatting();
+    private static final String RAW_PROTO_DOC_WITH_PRE_TAG =
+            "/** Doc header <pre> Preformated doc </pre> <code>string field = 1;</code> */";
+
+    private final Formatting formatting = new RemovePreTags();
 
     @Test
-    @DisplayName("merge lines")
-    void mergeLines() {
-        String lineText = "a text in a single line";
-        int lineCount = 5;
-        Iterable<String> lines = Collections.nCopies(lineCount, lineText);
-        String expectedLines = Joiner.on(lineSeparator())
-                                     .join(lines);
-
-        String formattedLines = formatting.apply(expectedLines);
-
-        assertThat(formattedLines)
-                .isEqualTo(expectedLines);
+    @DisplayName("remove tags generated by proto compiler")
+    void remove_tags_generated_by_proto_compiler() {
+        String result = formatting.apply(RAW_PROTO_DOC);
+        assertEquals(PROCESSED_PROTO_DOC, result);
     }
 
-    /**
-     * A stub formatting which simply returns the passed line.
-     */
-    private static class NothingFormatting extends LineFormatting {
+    @Test
+    @DisplayName("not remove user pre tags from doc")
+    void not_remove_user_pre_tags_from_doc() {
+        assertEquals(RAW_PROTO_DOC_WITH_PRE_TAG, formatting.apply(RAW_PROTO_DOC_WITH_PRE_TAG));
+    }
 
-        @Override
-        String formatLine(String line) {
-            return line;
-        }
+    @Test
+    @DisplayName("not format javadoc without pre tags")
+    void not_format_javadoc_without_pre_tags() {
+        String javadoc = "/** smth */";
+        assertEquals(javadoc, formatting.apply(javadoc));
     }
 }
