+++ tools/javadoc-style/src/test/java/io/spine/tools/javadoc/style/gradle/JavadocStylePluginFormattingTest.java
@@ -40,9 +40,9 @@ import java.nio.file.Path;
 import java.util.List;
 
 import static com.google.common.truth.Truth.assertThat;
-import static io.spine.tools.javadoc.style.formatting.BacktickFormatting.BACKTICK;
-import static io.spine.tools.javadoc.style.formatting.PreTagFormatting.CLOSING_PRE;
-import static io.spine.tools.javadoc.style.formatting.PreTagFormatting.OPENING_PRE;
+import static io.spine.tools.javadoc.style.formatting.BacktickedToCode.BACKTICK;
+import static io.spine.tools.javadoc.style.formatting.RemovePreTags.CLOSING_PRE;
+import static io.spine.tools.javadoc.style.formatting.RemovePreTags.OPENING_PRE;
 import static io.spine.tools.javadoc.style.gradle.JavadocStyleTaskName.formatProtoDoc;
 import static java.lang.System.lineSeparator;
 import static java.nio.charset.StandardCharsets.UTF_8;
